Add basic structure & authentication
The basic folder structure has been setout, but is subject to change.
In addition, authentication has now been completed, so that the
application can actually communicate with Spotify. The OAuth2 token
is stored in a file in plaintext though, seeing as this is a Hackweek
project and not missile flight code, so do be careful. An updated README
will follow shortly, with more details on the project and some
considerations to make.