Add player functionality
Adds player functionality and the barebones CLI interface for it.
Currently, this is done with the wmenu package, although this
could change to the promptui package, for aesthetic reasons.
In addition, authentication is now abstracted away, so it's up to
each component that needs authentication to obtain it/ensure it's already been done, rather than
having the main method first obtain authentication. This way,
a component simply needs to be created with something like
`playerControl.NewPlayer()` and authentication will be taken care of
during component creation.