Update file structure & switch to promptui
Promptui is now being used instead of wmenu, as it enables a more
feature rich CLI, in addition to being prettier. In addition,
the player and authentication files have been split into their own
modules, for maintainability. Finally, fixed the bug where
the user couldn't select a action to take after authenticating
for the first time.